refactor: move to utils handleSortableDragEnd
